Create new branch only if it doesn't exist, otherwise it errors
The git_commit agent task status would be true if everything
goes well the first time but on a repeated request it would raise
an exception that the branch already exists and the task_manager
would set the task status false as it saw an exception. So the
admin would get a report with git_commit agent status set to
false even though it ran as expected the first time.